공통 타입 분리 / word 관련기능 모듈 분리 / word 데이터 구조 변경
