+++ src/background/index.ts
@@ -1,10 +1,15 @@
-import { saveWord } from '@/utils/storage';
+import { saveWord, getMeaning } from '@/utils/word';
 
 global.browser = require('webextension-polyfill');
 
 chrome.runtime.onMessage.addListener(receiver);
 
 async function receiver(request: any, sender?: any, sendResponse?: any) {
-    saveWord(request.word);
+    const word: IWord = {
+        name: request.word,
+        meaning: await getMeaning(request.word)
+    };
+
+    saveWord(word);
 }
