+++ src/utils/storage.ts
@@ -1,6 +1,4 @@
 import { APP_NAME } from '@/constants/app';
-import { checkWordValid } from './format';
-import * as  moment from 'moment';
 
 export function saveData(key: string, data: any): Promise<any> {
     key = `${APP_NAME}_${key}`;
@@ -41,19 +39,3 @@ export function loadData(key: string): Promise<any> {
         }
     );
 }
-
-export async function saveWord(word: string, date: Date = new Date()): Promise<any> {
-    if (!word || !checkWordValid(word)) { return; }
-
-    const yyyymmdd = moment(new Date()).format('YYYYMMDD');
-    const key = `${yyyymmdd}`;
-    const savedData = await loadData(key);
-    const wordList = savedData || [];
-
-    console.log(wordList);
-
-    wordList.push(word);
-    console.log(wordList);
-
-    return saveData(key, wordList);
-};
