+++ tsconfig.json
@@ -10,7 +10,7 @@
         "moduleResolution": "node",
         "typeRoots": [
             "./node_modules/@types",
-            "./types"
+            "./src/types"
         ],
         "resolveJsonModule": true,
         "noImplicitAny": false,
