+++ types/app.d.ts
@@ -1,0 +1,11 @@
+interface IMenu {
+    id: string;
+    title: string;
+    path: string;
+    iconClassName?: string;
+}
+
+interface IWord {
+    name: string;
+    meaning: string;
+}
